[Main, Maunal]Topview cam 추가 및 영상 컷팅 미러 플립적용 & 로봇 대시보드 수정

+++ gui/src/lovo_gui/lovo_gui/tabs/tab02_manual/robot_arm/robot_dashboard.py
@@ -577,8 +577,8 @@ class RobotDashboardWidget(QWidget):
             pub = self.controller.get_publisher('order_command')
             
             if pub:
-                from std_msgs.msg import Float32
-                msg = Float32()
+                from std_msgs.msg import Float64
+                msg = Float64()
                 msg.data = value
                 pub.publish(msg)
